+++ src/step/content/replace/html/class/ClassDomReplaceCommand.ts
@@ -1,7 +1,6 @@
 import { ReplacerFactory } from "../../ReplacerFactory.js"
 import { DomReplaceCommand } from "../../DomReplaceCommand.js"
 import { DomReplacer } from "../../DomReplacer.js"
-import { HtmlSsgContext } from "../../../../../HtmlSsgContext.js"
 
 /**
  * A replacer that looks for HTML class(es) in tags.
@@ -12,11 +11,7 @@ export class ClassDomReplaceCommand<T extends HTMLElement = HTMLElement> extends
    * @param replacerFactory The object that creates the replacer when needed.
    * @param classNames The classname(s) to look for.
    */
-  constructor(protected replacerFactory: ReplacerFactory<DomReplacer<T>>, ...classNames: string[]) {
-    super(classNames.map(className => "." + className).join(","))
-  }
-
-  protected createReplacer(context: HtmlSsgContext): Promise<DomReplacer<T>> {
-    return this.replacerFactory.create(context)
+  constructor(replacerFactory: ReplacerFactory<DomReplacer<T>>, ...classNames: string[]) {
+    super(classNames.map(className => "." + className).join(","), replacerFactory)
   }
 }