+++ src/step/content/replace/html/tag/index.ts
@@ -1,2 +1,0 @@
-export * from "./RegexHtmlTagReplaceCommand.js"
-export * from "./HtmlTagReplaceCommand.js"